Extract the common namespace and schema as fixtures for the shared tests
