+++ crates/integration_tests/tests/shared_tests/read_positional_deletes.rs
